Fix bug causing get_mmio index overflow for swift
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83359
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Dev-Ready: Benjamin Gass <bgass@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83373
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>

+++ src/import/chips/p9/procedures/hwp/nest/p9a_put_mmio.C
@@ -5,7 +5,7 @@
 /*                                                                        */
 /* OpenPOWER HostBoot Project                                             */
 /*                                                                        */
-/* Contributors Listed Below - COPYRIGHT 2018                             */
+/* Contributors Listed Below - COPYRIGHT 2018,2019                        */
 /* [+] International Business Machines Corp.                              */
 /*                                                                        */
 /*                                                                        */
@@ -55,8 +55,8 @@ fapi2::ReturnCode p9a_put_mmio(const fapi2::Target<fapi2::TARGET_TYPE_OCMB_CHIP>
                                const std::vector<uint8_t>& i_data)
 {
     uint8_t data[8];
-    uint8_t l_idx;
-    uint8_t l_data_idx;
+    uint32_t l_idx;
+    uint32_t l_data_idx;
     uint32_t l_max_grans;
     uint32_t l_grans;
     p9_ADU_oper_flag l_myAduFlag;
